Email setup in /utils/email.js
I created an email.js file that holds the functions that refresh the tokens used to connect to our new matchpawfect@gmail.com email. We are also able to send emails, change the subject and the text accordingly as I created variables in place.  To test this code type: " node utils/email.js " into the terminal. Thanks :)