added logout button and signup email notificaiton
added logout button to main views, and added an email notification when a user signups to the app. I moved the code from the email.js folder I had created to test this out last week. I moved it into the userRoutes.js as this is where the new users are added to the database.